% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.r
\name{qwdap.eval}
\alias{qwdap.eval}
\title{Evaluation}
\usage{
qwdap.eval(series1, series2)
}
\arguments{
\item{series1}{The series1.}

\item{series2}{The series2.}
}
\value{
Three indicators, the Coefficient of Determination, Root Mean Squared Error and
Mean Absolute Error.
}
\description{
calculate the Coefficient of Determination, Root Mean Squared Error and the
Mean Absolute Error between two series.
}
\examples{
set.seed(1)
res.eval <- qwdap.eval(rnorm(100,0,2),rnorm(100,0,1))
}
