% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.r
\name{qwdap.pcr}
\alias{qwdap.pcr}
\title{Principle Component Regression}
\usage{
qwdap.pcr(in_data, data_range, plotting)
}
\arguments{
\item{in_data}{a 'QWMS' object, which includes the target series and the
selected modes which can be obtained from modes selection.}

\item{data_range}{the range of the train samples.}

\item{plotting}{whether to plot.}
}
\value{
a 'QWMODEL' object which includes the information of regression analysis.
}
\description{
Principle component regression. This is a linear regression method used to
establish the linear relationship between the original time series and the modes
generated by quantum walks.
}
\examples{
data("traffic.n1")
res.pcr <- qwdap.pcr(traffic.n1,c(1,500), FALSE)

}
