% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr.r
\name{qwdap.ppr}
\alias{qwdap.ppr}
\title{Projection Pursuit Regression}
\usage{
qwdap.ppr(in_data, data_range, plotting)
}
\arguments{
\item{in_data}{a 'QWMS' object, which includes the target series and the
selected modes which can be obtained from modes selection.}

\item{data_range}{the range of the train samples.}

\item{plotting}{whether to plot.}
}
\value{
a 'QWMODEL' object which includes the information of regression analysis.
}
\description{
Projection pursuit regression. This is a nonlinear regression method used to
establish the nonlinear relationship between the original time series and the modes
generated by quantum walks.
}
\examples{
data("traffic.n1")
res.ppr <- qwdap.ppr(traffic.n1,c(1,500))

}
