% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{qwdap.predict}
\alias{qwdap.predict}
\title{Prediction}
\usage{
qwdap.predict(in_model, data_range)
}
\arguments{
\item{in_model}{a 'QWMODEL' object, which is the model built by Stepwise Regression,
PCR, PLSR, PPR, VAR in this package.}

\item{data_range}{indicate the index range of the part data generated by quantum walks for predict.}
}
\value{
the predict data.
}
\description{
Based on the established model, make predict. The core algorithm of VAR prediction comes from MTS(ver. 1.1.1).
}
\examples{
data(traffic.model.n1)
res.predict <- qwdap.predict(traffic.model.n1,c(501,720))
}
