% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrelieff.r
\name{qwdap.rrelieff}
\alias{qwdap.rrelieff}
\title{RReliefF}
\usage{
qwdap.rrelieff(real, ctqw, index, num, plotting)
}
\arguments{
\item{real}{the real series observed.}

\item{ctqw}{the 'CTQW' object.}

\item{index}{the index of the data for mode selection.}

\item{num}{the number of series required.}

\item{plotting}{whether to plot.}
}
\value{
a 'QWMS' object.
}
\description{
Mode selection by RReliefF. The purpose of this function
is to select the part modes with similar characteristics to the observed time series
from the modes generated by the quantum walk. And it is based on the data model.
}
\details{
The 'QWMS' object include the original time series and the modes generated by
quantum walks.
}
\examples{
data("traffic.qw")
data("trafficflow")
res.rrelieff <- qwdap.rrelieff(trafficflow,traffic.qw,1,30,TRUE)

}
