% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.r
\name{qwdap.var}
\alias{qwdap.var}
\title{Vector Autoregressive Model}
\usage{
qwdap.var(in_data, data_range, plotting)
}
\arguments{
\item{in_data}{a 'QWMS' object, which includes the target series and the
selected modes which can be obtained from modes selection.}

\item{data_range}{the range of the train samples.}

\item{plotting}{whether to plot.}
}
\value{
a 'QWMODEL' object which includes the information of regression analysis.
}
\description{
Vector autoregressive model. This is a regression method used to
establish the temporal relationship between the original time series and
the modes generated by quantum walks. The core algorithm comes from MTS(ver. 1.1.1).
}
\examples{
data("traffic.n1")
res.var <- qwdap.var(traffic.n1,c(1,500))

}
