\name{traffic.model.n1}
\docType{data}
\alias{traffic.model.n1}
\title{The estabulished model by Stepwise Regression of the 'N1' station}
\description{
  This data is the linear model built by Stepwise Regression of the highway traffic flow data of the 'N1' stations, and includes the observed data and the modes generated by quantum walk. 
}
\usage{data(traffic.model.n1)}
\format{A 'QWMODEL' object.}
\source{Pan BH(2021).}
\keyword{datasets}