\name{traffic.n1}
\docType{data}
\alias{traffic.n1}
\title{Data of the 'N1' station}
\description{
  This data is the highway traffic flow data and the modes generated by quantum walk of the 'N1' stations.
}
\usage{data(traffic.n1)}
\format{A 'QWMS' object.}
\source{Pan BH(2021).}
\keyword{datasets}