\name{traffic.qw}
\docType{data}
\alias{traffic.qw}
\title{A set of modes generated by quantum walk}
\description{
  This data is generated by function 'qwdap.qwalk()' with 100 scaling factors form 0.01 to 1 and the parameter 'edges' is the adjancy matrix of 7 vertices connected end to end.
}
\usage{data(traffic.qw)}
\format{A 'CTQW' object.}
\source{Pan BH(2021).}
\keyword{datasets}