\name{trafficflow}
\docType{data}
\alias{trafficflow}
\title{Highway traffic flow data}
\description{
  This data set has a total of 720 records of 7 research stations, namely Tangshan (N1), Jurong (N2), Heyang (N3), Danyang (N4), Luoshuyan (N5), Xuejia (N6) and ChangzhouBei (N7).
}
\usage{data(trafficflow)}
\format{A dataframe with 720 observations on the 7 stations.}
\source{Yu ZY, Hu X(2020).}
\keyword{datasets}